requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.0
Django==4.2.8
django-celery-beat==2.5.0
django-cors-headers==4.2.0
django-filter==23.2
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
inflection==0.5.1
install==1.3.5
kombu==5.3.4
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
psycopg==3.1.14
PyJWT==2.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.12
